LayoutTests/fast/frames/seamless/seamless-inherited-origin-expected.txt

Test that frames with inherited security origins can still be seamless.
FAIL iframe.seamless should be true (of type boolean). Was undefined (of type undefined).
PASS window.getComputedStyle(iframe).width is "200px"
PASS window.getComputedStyle(iframe).height is "8px"
PASS window.getComputedStyle(iframe).width is "200px"
PASS window.getComputedStyle(iframe).height is "100px"
